// src/utils/api/api.ts
export const ActivityLevel = {
    Weak: 'WEAK',
    Growing: 'GROWING',
    Strong: 'STRONG',
    Restricted: 'RESTRICTED'
} as const;

export type ActivityLevel = typeof ActivityLevel[keyof typeof ActivityLevel];
// ...
export interface ShipEngine {
    /**
     * The symbol of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'symbol': ShipEngineSymbolEnum;
    /**
     * The name of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'name': string;
    /**
     * The description of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipEngine
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipEngine
     */
    'integrity': number;
    /**
     * The speed stat of this engine. The higher the speed, the faster a ship can travel from one point to another. Reduces the time of arrival when navigating the ship.
     * @type {number}
     * @memberof ShipEngine
     */
    'speed': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipEngine
     */
    'requirements': ShipRequirements;
}
// ...
export interface ShipFrame {
    /**
     * Symbol of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'symbol': ShipFrameSymbolEnum;
    /**
     * Name of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'name': string;
    /**
     * Description of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipFrame
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipFrame
     */
    'integrity': number;
    /**
     * The amount of slots that can be dedicated to modules installed in the ship. Each installed module take up a number of slots, and once there are no more slots, no new modules can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'moduleSlots': number;
    /**
     * The amount of slots that can be dedicated to mounts installed in the ship. Each installed mount takes up a number of points, and once there are no more points remaining, no new mounts can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'mountingPoints': number;
    /**
     * The maximum amount of fuel that can be stored in this ship. When refueling, the ship will be refueled to this amount.
     * @type {number}
     * @memberof ShipFrame
     */
    'fuelCapacity': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipFrame
     */
    'requirements': ShipRequirements;
}
// ...
export interface ShipModule {
    /**
     * The symbol of the module.
     * @type {string}
     * @memberof ShipModule
     */
    'symbol': ShipModuleSymbolEnum;
    /**
     * Modules that provide capacity, such as cargo hold or crew quarters will show this value to denote how much of a bonus the module grants.
     * @type {number}
     * @memberof ShipModule
     */
    'capacity'?: number;
    /**
     * Modules that have a range will such as a sensor array show this value to denote how far can the module reach with its capabilities.
     * @type {number}
     * @memberof ShipModule
     */
    'range'?: number;
    /**
     * Name of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'name': string;
    /**
     * Description of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'description': string;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipModule
     */
    'requirements': ShipRequirements;
}
// ...
export interface ShipMount {
    /**
     * Symbo of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'symbol': ShipMountSymbolEnum;
    /**
     * Name of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'name': string;
    /**
     * Description of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'description'?: string;
    /**
     * Mounts that have this value, such as mining lasers, denote how powerful this mount\'s capabilities are.
     * @type {number}
     * @memberof ShipMount
     */
    'strength'?: number;
    /**
     * Mounts that have this value denote what goods can be produced from using the mount.
     * @type {Array<string>}
     * @memberof ShipMount
     */
    'deposits'?: Array<ShipMountDepositsEnum>;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipMount
     */
    'requirements': ShipRequirements;
}
// ...
export interface ShipReactor {
    /**
     * Symbol of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'symbol': ShipReactorSymbolEnum;
    /**
     * Name of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'name': string;
    /**
     * Description of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipReactor
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipReactor
     */
    'integrity': number;
    /**
     * The amount of power provided by this reactor. The more power a reactor provides to the ship, the lower the cooldown it gets when using a module or mount that taxes the ship\'s power.
     * @type {number}
     * @memberof ShipReactor
     */
    'powerOutput': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipReactor
     */
    'requirements': ShipRequirements;
}
// ...
export const ShipType = {
    Probe: 'SHIP_PROBE',
    MiningDrone: 'SHIP_MINING_DRONE',
    SiphonDrone: 'SHIP_SIPHON_DRONE',
    Interceptor: 'SHIP_INTERCEPTOR',
    LightHauler: 'SHIP_LIGHT_HAULER',
    CommandFrigate: 'SHIP_COMMAND_FRIGATE',
    Explorer: 'SHIP_EXPLORER',
    HeavyFreighter: 'SHIP_HEAVY_FREIGHTER',
    LightShuttle: 'SHIP_LIGHT_SHUTTLE',
    OreHound: 'SHIP_ORE_HOUND',
    RefiningFreighter: 'SHIP_REFINING_FREIGHTER',
    Surveyor: 'SHIP_SURVEYOR'
} as const;

export type ShipType = typeof ShipType[keyof typeof ShipType];


/**
 * 
 * @export
 * @interface Shipyard
 */
export interface Shipyard {
    /**
     * The symbol of the shipyard. The symbol is the same as the waypoint where the shipyard is located.
     * @type {string}
     * @memberof Shipyard
     */
    'symbol': string;
    /**
     * The list of ship types available for purchase at this shipyard.
     * @type {Array<ShipyardShipTypesInner>}
     * @memberof Shipyard
     */
    'shipTypes': Array<ShipyardShipTypesInner>;
    /**
     * The list of recent transactions at this shipyard.
     * @type {Array<ShipyardTransaction>}
     * @memberof Shipyard
     */
    'transactions'?: Array<ShipyardTransaction>;
    /**
     * The ships that are currently available for purchase at the shipyard.
     * @type {Array<ShipyardShip>}
     * @memberof Shipyard
     */
    'ships'?: Array<ShipyardShip>;
    /**
     * The fee to modify a ship at this shipyard. This includes installing or removing modules and mounts on a ship. In the case of mounts, the fee is a flat rate per mount. In the case of modules, the fee is per slot the module occupies.
     * @type {number}
     * @memberof Shipyard
     */
    'modificationsFee': number;
}
/**
 * 
 * @export
 * @interface ShipyardShip
 */
export interface ShipyardShip {
    /**
     * 
     * @type {ShipType}
     * @memberof ShipyardShip
     */
    'type': ShipType;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'description': string;
    /**
     * 
     * @type {SupplyLevel}
     * @memberof ShipyardShip
     */
    'supply': SupplyLevel;
    /**
     * 
     * @type {ActivityLevel}
     * @memberof ShipyardShip
     */
    'activity'?: ActivityLevel;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShip
     */
    'purchasePrice': number;
    /**
     * 
     * @type {ShipFrame}
     * @memberof ShipyardShip
     */
    'frame': ShipFrame;
    /**
     * 
     * @type {ShipReactor}
     * @memberof ShipyardShip
     */
    'reactor': ShipReactor;
    /**
     * 
     * @type {ShipEngine}
     * @memberof ShipyardShip
     */
    'engine': ShipEngine;
    /**
     * 
     * @type {Array<ShipModule>}
     * @memberof ShipyardShip
     */
    'modules': Array<ShipModule>;
    /**
     * 
     * @type {Array<ShipMount>}
     * @memberof ShipyardShip
     */
    'mounts': Array<ShipMount>;
    /**
     * 
     * @type {ShipyardShipCrew}
     * @memberof ShipyardShip
     */
    'crew': ShipyardShipCrew;
}


/**
 * 
 * @export
 * @interface ShipyardShipCrew
 */
export interface ShipyardShipCrew {
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'required': number;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'capacity': number;
}
/**
 * 
 * @export
 * @interface ShipyardShipTypesInner
 */
export interface ShipyardShipTypesInner {
    /**
     * 
     * @type {ShipType}
     * @memberof ShipyardShipTypesInner
     */
    'type': ShipType;
}


/**
 * Results of a transaction with a shipyard.
 * @export
 * @interface ShipyardTransaction
 */
export interface ShipyardTransaction {
    /**
     * The symbol of the waypoint.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'waypointSymbol': string;
    /**
     * The symbol of the ship that was the subject of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     * @deprecated
     */
    'shipSymbol': string;
    /**
     * The symbol of the ship that was the subject of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'shipType': string;
    /**
     * The price of the transaction.
     * @type {number}
     * @memberof ShipyardTransaction
     */
    'price': number;
    /**
     * The symbol of the agent that made the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'agentSymbol': string;
    /**
     * The timestamp of the transaction.
     * @type {string}
     * @memberof ShipyardTransaction
     */
    'timestamp': string;
}
// ...
export const SupplyLevel = {
    Scarce: 'SCARCE',
    Limited: 'LIMITED',
    Moderate: 'MODERATE',
    High: 'HIGH',
    Abundant: 'ABUNDANT'
} as const;

export type SupplyLevel = typeof SupplyLevel[keyof typeof SupplyLevel];